Promoting JSON ostream to shared
