+++ tools/test-tools/qa-stats/CMakeLists.txt
@@ -5,14 +5,13 @@ project(qa-stats)
 set(SOURCES 
     main.cpp
     input.cpp
-    output.cpp
 )
 
 set(HEADERS
     input.hpp
     hashing.hpp
-    output.hpp
     parameters.hpp
+    stats.hpp
 )
 
 if (NOT WIN32)
@@ -21,6 +20,11 @@ if (NOT WIN32)
     add_executable(qa-stats ${SOURCES} ${HEADERS})
     target_compile_features(qa-stats PRIVATE cxx_std_17)
     target_link_libraries(qa-stats Threads::Threads)
+    
+#     target_compile_options( qa-stats PRIVATE -fsanitize=undefined )
+#     target_compile_options( qa-stats PRIVATE -fsanitize=address )
+#     target_link_options( qa-stats PRIVATE -fsanitize=undefined )
+#     target_link_options( qa-stats PRIVATE -fsanitize=address )
 endif()
 
 source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" PREFIX "Source Files" FILES ${SOURCES})
