+++ tools/test-tools/qa-stats/main.cpp
@@ -37,11 +37,196 @@
 #include <tuple>
 #include <cmath>
 #include <cassert>
+#include <JSON_ostream.hpp>
 #include "parameters.hpp"
 #include "input.hpp"
 #include "hashing.hpp"
 #include "stats.hpp"
-#include "output.hpp"
+
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &s, HashResult64 const &v) {
+    char buffer[32];
+    return s << HashResult64::to_hex(v.value, buffer);
+}
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, CountBT const &value) {
+    return out
+        << JSON_Member{"count"} << value.total
+        << JSON_Member{"biological"} << value.biological
+        << JSON_Member{"technical"} << value.technical;
+}
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, CountFR const &value) {
+    return out
+        << JSON_Member{"total"} << value.total
+        << JSON_Member{"5'"} << value.fwd
+        << JSON_Member{"3'"} << value.rev;
+}
+
+struct BaseStatsEntry {
+    char const *base;
+    uint64_t bio, tech;
+
+    friend
+    JSON_ostream &operator <<(JSON_ostream &out, BaseStatsEntry const &self) {
+        if (self.bio || self.tech) {
+            out << '{' << JSON_Member{"base"} << self.base;
+            if (self.bio)
+                out << JSON_Member{"biological"} << self.bio;
+            if (self.tech)
+                out << JSON_Member{"technical"} << self.tech;
+            out << '}';
+        }
+        return out;
+    }
+};
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, BaseStats const &stats) {
+    using Index = BaseStats::Index;
+    auto const bioW = stats[Index{'A', false}] + stats[Index{'T', false}];
+    auto const bioS = stats[Index{'C', false}] + stats[Index{'G', false}];
+    auto const bioN = stats[Index{'N', false}];
+    auto const techW = stats[Index{'A', true}] + stats[Index{'T', true}];
+    auto const techS = stats[Index{'C', true}] + stats[Index{'G', true}];
+    auto const techN = stats[Index{'N', true}];
+
+    return out
+        << BaseStatsEntry{"N", bioN, techN}
+        << BaseStatsEntry{"W", bioW, techW}
+        << BaseStatsEntry{"S", bioS, techS};
+}
+
+struct DistanceStatEntry {
+    double power;
+    unsigned length;
+
+    DistanceStatEntry(DistanceStats::DistanceStat::Index i, uint64_t count, uint64_t total)
+    : power(double(count)/total)
+    , length(i)
+    {}
+
+    friend
+    JSON_ostream &operator <<(JSON_ostream &out, DistanceStatEntry const &self) {
+        return out
+        << '{'
+            << JSON_Member{"length"} << self.length
+            << JSON_Member{"power"} << self.power
+        << '}';
+    }
+};
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, DistanceStats::DistanceStat const &self) {
+    using Index = DistanceStats::DistanceStat::Index;
+    self.forEach([&](Index i, uint64_t count, uint64_t total) {
+        out << DistanceStatEntry{ i, count, total };
+    });
+    return out;
+}
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, std::pair < DistanceStats::DistanceStat const *, DistanceStats::DistanceStat const * > const &self)
+{
+    using Index = DistanceStats::DistanceStat::Index;
+    self.first->forEach(*self.second, [&](Index i, uint64_t count, uint64_t total) {
+        out << DistanceStatEntry{ i, count, total };
+    });
+    return out;
+}
+
+static inline
+JSON_ostream &operator <<(JSON_ostream &out, DistanceStats const &distance) {
+    if (distance) {
+        out << JSON_Member{"W"} << '[' << std::make_pair(&distance.A, &distance.T) << ']';
+        out << JSON_Member{"S"} << '[' << std::make_pair(&distance.C, &distance.G) << ']';
+        out << JSON_Member{"S-W"} << '[' << distance.SW << ']';
+        out << JSON_Member{"K-M"} << '[' << distance.KM << ']';
+    }
+    return out;
+}
+
+static
+JSON_ostream &operator <<(JSON_ostream &out, ReferenceStats const &self) {
+    for (auto &reference : self) {
+        auto const refId = &reference - &self[0];
+        auto const refName = Input::references[refId];
+        CountFR total;
+        auto const &posHash = reference[0].total;
+        auto const &seqHash = reference[0].fwd;
+        auto const &cigHash = reference[0].rev;
+
+        out << '{'
+            << JSON_Member{"name"} << refName
+            << JSON_Member{"chunks"} << '[';
+        for (auto &counter : reference) {
+            if (&counter == &reference[0] || counter.total == 0) continue;
+
+            auto const chunk = (&counter - &reference[0]) - 1;
+            total += counter;
+
+            out
+            << '{'
+                << JSON_Member{"start"} << chunk * ReferenceStats::chunkSize + 1
+                << JSON_Member{"last"} << (chunk + 1) * ReferenceStats::chunkSize
+                << JSON_Member{"alignments"} << '{' << counter << '}'
+            << '}';
+        }
+        out << ']'
+            << JSON_Member{"alignments"} << '{' << total << '}'
+            << JSON_Member{"position"} << HashResult64{posHash}
+            << JSON_Member{"sequence"} << HashResult64{seqHash}
+            << JSON_Member{"cigar"} << HashResult64{cigHash}
+        << '}';
+    }
+    return out;
+}
+
+static JSON_ostream &operator <<(JSON_ostream &out, SpotLayouts const &value) {
+    for (auto i = value.index.begin(); i != value.index.end(); ++i) {
+        out << '{'
+            << JSON_Member{"descriptor"} << i->first
+            << JSON_Member{"count"} << value.count[i->second]
+        << '}';
+    }
+    return out;
+}
+
+static JSON_ostream &operator <<(JSON_ostream &out, CountBTN const &value) {
+    return out
+        << JSON_Member{"total"} << value.total
+        << JSON_Member{"biological"} << value.biological
+        << JSON_Member{"no-biological"} << value.nobiological
+        << JSON_Member{"technical"} << value.technical
+        << JSON_Member{"no-technical"} << value.notechnical;
+}
+
+static
+JSON_ostream &operator <<(JSON_ostream &out, SpotStats const &stats) {
+    for (auto &[k, v] : stats) {
+        out << '{'
+            << JSON_Member{"nreads"} << k
+            << v
+        << '}';
+    }
+    return out;
+}
+
+static JSON_ostream &operator <<(JSON_ostream &out, Stats const &self) {
+    out << JSON_Member{"reads"}        << '{' << self.reads      << '}';
+    out << JSON_Member{"bases"}        << '[' << self.bases      << ']';
+    out << JSON_Member{"spots"}        << '[' << self.spots      << ']';
+    out << JSON_Member{"spot-layouts"} << '[' << self.layouts    << ']';
+    out << JSON_Member{"spectra"}      << '{' << self.spectra    << '}';
+    out << JSON_Member{"references"}   << '[' << self.references << ']';
+    // fingerprint is only output separately
+    // out << JSON_Member{"fingerprint"}  << '[' << self.fingerprint << ']';
+
+    return out;
+}
 
 static
 SpotLayout spotLayout(Input const &spot) {
@@ -283,7 +468,7 @@ int main(int argc, char * argv[]) {
     Input::runTests();
     SeqHash::test();
  */
-    auto app = App(argc, argv);
+    auto app = App{argc, argv};
 
     return app.run();
 }
